requirements.txt
asgiref==3.11.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.0.6
djangorestframework==3.16.1
mysqlclient==2.2.7
numpy==2.4.1
pandas==3.0.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pycparser==2.23
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
